--- .github/actions/deploy/action.yml
@@ -1,32 +1,0 @@
-name: 'Deploy'
-inputs:
-  model_name: 
-    description: 'Model Name'
-    required: true
-  tenant_id: 
-    description: 'Tenant Id'
-    required: true
-  service_principal_id: 
-    description: 'Service Principal Id'
-    required: true
-  service_principal_pwd: 
-    description: 'Service Principal Pwd'
-    required: true
-  subscription_id: 
-    description: 'Subscription Id'
-    required: true
-  resource_group: 
-    description: 'Resource Group'
-    required: true
-  workspace: 
-    description: 'Workspace'
-    required: true
-  run_id: 
-    description: 'Run Id'
-    required: true
-runs:
-  using: 'docker'
-  image: 'eugenefedorenko/mldeploy'
-  args:
-    - "sh"
-    - "/scripts/deploy.sh"
