+++ .github/events/event_dispatcher.py
@@ -1,12 +1,14 @@
 import argparse
-from dispatch_event_handler import DispatchEventHandler
+from dispatch_event_handler import BaseDispatchEventHandler
 from reg_model_event_handler import RegModelEventHandler
 from start_train_event_handler import StartEventHandler
 from finish_train_event_handler import FinishEventHandler
 
+
 REGISTER_MODEL_EVENT = "Model is registered"
 START_TRAIN_EVENT = "Training Started"
 FINISH_TRAIN_EVENT = "Training Finished"
+
 
 def get_event_handler(event_type):
     if (event_type == REGISTER_MODEL_EVENT):
@@ -16,7 +18,7 @@ def get_event_handler(event_type):
     elif (event_type == FINISH_TRAIN_EVENT):
         return FinishEventHandler()
     else:
-        return DispatchEventHandler()
+        return BaseDispatchEventHandler()
 
 
 if __name__ == "__main__":
@@ -26,4 +28,4 @@ if __name__ == "__main__":
     args = parser.parse_args()
 
     event_dispatcher = get_event_handler(args.event_type)
-    event_dispatcher.dispatch() 
+    event_dispatcher.dispatch()
