--- actions_utils/kubemlopsbot/ghcreds-secret.sh
@@ -1,1 +1,0 @@
-kubectl create secret generic ghcreds-secret --from-literal=GITHUB_TOKEN='3f2579a237fc3bdcec270737f485e7d12dffdb01' --from-literal=GITHUB_REPOSITORY='eedorenko/kubemlops' -n NAMESPACE
